Fixed #1783 catalog UI review
