+++ web/client/components/catalog/Catalog.jsx
@@ -1,5 +1,4 @@
-const PropTypes = require('prop-types');
-/**
+/*
  * Copyright 2017, GeoSolutions Sas.
  * All rights reserved.
  *
@@ -7,11 +6,14 @@ const PropTypes = require('prop-types');
  * LICENSE file in the root directory of this source tree.
  */
 const React = require('react');
-
+const PropTypes = require('prop-types');
 const Message = require('../I18N/Message');
 const LocaleUtils = require('../../utils/LocaleUtils');
+const ComboBoxListItem = require('../misc/ComboBoxListItem');
+const ComboboxReact = require('react-widgets').Combobox;
+const assign = require('object-assign');
 
-const {FormControl, FormGroup, Alert, Pagination, Button, Panel} = require('react-bootstrap');
+const {FormControl, FormGroup, Alert, Pagination, Button, Panel, Form, Col, InputGroup, ControlLabel, Glyphicon} = require('react-bootstrap');
 const Spinner = require('react-spinkit');
 
 const RecordGrid = require('./RecordGrid');
@@ -19,33 +21,48 @@ const RecordGrid = require('./RecordGrid');
 class Catalog extends React.Component {
     static propTypes = {
         active: PropTypes.bool,
+        addAuthentication: PropTypes.bool,
+        buttonClassName: PropTypes.string,
+        buttonStyle: PropTypes.object,
+        chooseCatalogUrl: PropTypes.bool,
+        displayURL: PropTypes.bool,
         formats: PropTypes.array,
         format: PropTypes.string,
-        searchOnStartup: PropTypes.bool,
-        onSearch: PropTypes.func,
-        onReset: PropTypes.func,
-        onChangeFormat: PropTypes.func,
-        onLayerAdd: PropTypes.func,
-        onZoomToExtent: PropTypes.func,
-        zoomToLayer: PropTypes.bool,
-        onError: PropTypes.func,
-        pageSize: PropTypes.number,
-        displayURL: PropTypes.bool,
-        initialCatalogURL: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
-        result: PropTypes.object,
-        loadingError: PropTypes.object,
-        layerError: PropTypes.string,
-        searchOptions: PropTypes.object,
-        chooseCatalogUrl: PropTypes.bool,
-        showGetCapLinks: PropTypes.bool,
-        addAuthentication: PropTypes.bool,
-        records: PropTypes.array,
         gridOptions: PropTypes.object,
+        initialCatalogURL: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
+        includeAddServiceButton: PropTypes.bool,
+        includeBackToViewButton: PropTypes.bool,
         includeSearchButton: PropTypes.bool,
         includeResetButton: PropTypes.bool,
+        loadingError: PropTypes.object,
+        layerError: PropTypes.string,
+        mode: PropTypes.string,
+        newService: PropTypes.object,
+        onAddService: PropTypes.func,
+        onChangeCatalogMode: PropTypes.func,
+        onChangeFormat: PropTypes.func,
+        onChangeNewTitle: PropTypes.func,
+        onChangeNewType: PropTypes.func,
+        onChangeNewUrl: PropTypes.func,
+        onChangeSelectedService: PropTypes.func,
+        onError: PropTypes.func,
+        onFocusServicesList: PropTypes.func,
+        onLayerAdd: PropTypes.func,
+        onReset: PropTypes.func,
+        onSearch: PropTypes.func,
+        onZoomToExtent: PropTypes.func,
+        openCatalogServiceList: PropTypes.bool,
+        pageSize: PropTypes.number,
+        records: PropTypes.array,
+        result: PropTypes.object,
+        searchOnStartup: PropTypes.bool,
+        searchOptions: PropTypes.object,
+        selectedService: PropTypes.string,
+        services: PropTypes.object,
+        serviceComboboxMessages: PropTypes.object,
+        showGetCapLinks: PropTypes.bool,
         wrapOptions: PropTypes.bool,
-        buttonStyle: PropTypes.object,
-        buttonClassName: PropTypes.string
+        zoomToLayer: PropTypes.bool
     };
 
     static contextTypes = {
@@ -53,25 +70,48 @@ class Catalog extends React.Component {
     };
 
     static defaultProps = {
-        pageSize: 6,
-        onSearch: () => {},
-        onReset: () => {},
-        onChangeFormat: () => {},
-        onLayerAdd: () => {},
-        onZoomToExtent: () => {},
-        zoomToLayer: true,
-        onError: () => {},
+        buttonClassName: "search-button",
+        buttonStyle: {
+            marginBottom: "10px",
+            marginRight: "5px"
+        },
         chooseCatalogUrl: true,
-        records: [],
         formats: [{name: 'csw', label: 'CSW'}],
         format: 'csw',
         includeSearchButton: true,
         includeResetButton: false,
-        wrapOptions: false,
-        buttonStyle: {
-            marginBottom: "10px"
+        includeBackToViewButton: true,
+        includeAddServiceButton: true,
+        mode: "view",
+        newService: {
+            title: "",
+            type: "wms",
+            url: ""
         },
-        buttonClassName: "search-button"
+        onAddService: () => {},
+        onChangeCatalogMode: () => {},
+        onChangeFormat: () => {},
+        onChangeNewTitle: () => {},
+        onChangeNewType: () => {},
+        onChangeNewUrl: () => {},
+        onChangeSelectedService: () => {},
+        onError: () => {},
+        onFocusServicesList: () => {},
+        onLayerAdd: () => {},
+        onReset: () => {},
+        onSearch: () => {},
+        onZoomToExtent: () => {},
+        openCatalogServiceList: false,
+        pageSize: 4,
+        records: [],
+        services: {},
+        serviceComboboxMessages: {
+            emptyList: <Message msgId="queryform.attributefilter.autocomplete.emptyList"/>,
+            open: <Message msgId="queryform.attributefilter.autocomplete.open"/>,
+            emptyFilter: <Message msgId="queryform.attributefilter.autocomplete.emptyFilter"/>
+        },
+        wrapOptions: false,
+        zoomToLayer: true
     };
 
     state = {
@@ -139,7 +179,7 @@ class Catalog extends React.Component {
         let pageSize = this.props.pageSize;
         let page = Math.floor( start / pageSize);
         let pageN = Math.ceil(total / pageSize);
-        return (<div><Pagination
+        return (<div style={{textAlign: "center"}}><Pagination
           prev next first last ellipsis boundaryLinks
           bsSize="small"
           items={pageN}
@@ -154,7 +194,7 @@ class Catalog extends React.Component {
     };
 
     renderRecords = () => {
-        return (<div>
+        return (<div className="catalog-results">
                 <RecordGrid {...this.props.gridOptions} key="records"
                     records={this.props.records}
                     catalogURL={this.getCatalogUrl() }
@@ -171,10 +211,11 @@ class Catalog extends React.Component {
 
     renderURLInput = () => {
         if (!this.getCatalogUrl() || this.props.chooseCatalogUrl) {
-            return (<FormGroup><FormControl
+            return (<FormGroup>
+                <FormControl
                 ref="catalogURL"
                 type="text"
-                placeholder={LocaleUtils.getMessageById(this.context.messages, "catalog.catalogUrlPlaceholder")}
+                placeholder={LocaleUtils.getMessageById(this.context.messages, "catalog.urlPlaceholder")}
                 onChange={this.setCatalogUrl}
                 onKeyDown={this.onKeyDown}/></FormGroup>);
         }
@@ -182,15 +223,25 @@ class Catalog extends React.Component {
 
     renderButtons = () => {
         const buttons = [];
-        if (this.props.includeSearchButton) {
-            buttons.push(<Button bsStyle="primary" style={this.props.buttonStyle} onClick={this.search}
-                        className={this.props.buttonClassName} key="catalog_search_button">
-                        {this.renderLoading()} <Message msgId="catalog.search"/>
-                    </Button>);
+        if (this.props.mode === "view" ) {
+            if (this.props.includeSearchButton) {
+                buttons.push(<Button bsStyle="primary" style={this.props.buttonStyle} onClick={this.search}
+                            className={this.props.buttonClassName} key="catalog_search_button" disabled={!this.isValidServiceSelected()}>
+                            {this.renderLoading()} <Message msgId="catalog.search"/>
+                        </Button>);
+            }
+            if (this.props.includeResetButton) {
+                buttons.push(<Button style={this.props.buttonStyle} onClick={this.reset} key="catalog_reset_button">
+                            <Message msgId="catalog.reset"/>
+                        </Button>);
+            }
         }
-        if (this.props.includeResetButton) {
-            buttons.push(<Button style={this.props.buttonStyle} onClick={this.reset} key="catalog_reset_button">
-                        <Message msgId="catalog.reset"/>
+        if (this.props.mode === "edit") {
+            buttons.push(<Button style={this.props.buttonStyle} onClick={() => this.props.onAddService()} key="catalog_add_service_button">
+                        <Message msgId="save"/>
+                    </Button>);
+            buttons.push(<Button style={this.props.buttonStyle} onClick={() => this.props.onChangeCatalogMode("view")} key="catalog_back_view_button">
+                        <Message msgId="cancel"/>
                     </Button>);
         }
         return buttons;
@@ -203,12 +254,8 @@ class Catalog extends React.Component {
         return null;
     };
 
-    renderFormats = () => {
-        return this.props.formats.map((format) => <option value={format.name} key={format.name}>{format.label}</option>);
-    };
-
-    render() {
-        const textSearch = (<FormGroup><FormControl
+    renderTextSearch = () => {
+        const textSearch = (<FormControl
             ref="searchText"
             type="text"
             style={{
@@ -216,32 +263,148 @@ class Catalog extends React.Component {
             }}
             placeholder={LocaleUtils.getMessageById(this.context.messages, "catalog.textSearchPlaceholder")}
             onChange={this.onSearchTextChange}
-            onKeyDown={this.onKeyDown}/></FormGroup>);
+            onKeyDown={this.onKeyDown}/>);
+        return this.props.wrapOptions ? (<Panel collapsible defaultExpanded={false} header={LocaleUtils.getMessageById(this.context.messages, "catalog.options")}>
+            {textSearch}
+        </Panel>) : textSearch;
+    }
+    renderFormats = () => {
+        return this.props.formats.map((format) => <option value={format.name} key={format.name}>{format.label}</option>);
+    };
+    renderServices = () => {
+        return Object.keys(this.props.services).map((service) => {
+            return service === this.props.selectedService ?
+            <option value={service} selected key={service}>{service}</option>
+            : <option value={service} key={service}>{service}</option>;
+        });
+    };
+    render() {
+        const services = Object.keys(this.props.services).map(s => {
+            return assign({}, this.props.services[s], {label: this.props.services[s].title});
+        });
         return (
-             <div>
-                 <div>
-                     {this.renderFormatChoice()}
-                     {this.renderURLInput()}
+            <div>
+                <div>
+                {this.isViewMode(this.props.mode) ? (
+                    <Form>
+                        <FormGroup controlId="labelService" key="labelService">
+                            <ControlLabel><Message msgId="catalog.service"/></ControlLabel>
+                        </FormGroup>
+                        <FormGroup controlId="service" key="service">
+                            <InputGroup>
+                                <ComboboxReact
+                                    data={services}
+                                    messages={this.props.serviceComboboxMessages}
+                                    value={this.props.selectedService}
+                                    open={this.props.openCatalogServiceList}
+                                    onFocus={(e) => {
+                                        if (!(e.target instanceof Node && e.target.nodeName === 'BUTTON')) {
+                                            this.props.onFocusServicesList(true);
+                                        }
+                                    }}
+                                    onToggle={() => this.props.onFocusServicesList(!this.props.openCatalogServiceList)}
+                                    onBlur={(e) => {
+                                        if (!this.props.services[e.target.value]) {
+                                            this.props.onChangeSelectedService("");
+                                        }
+                                    }}
+                                    onChange={(val) => this.props.onChangeSelectedService(val.title ? val.title : val)}
+                                    onSelect={(service) => this.props.onChangeSelectedService(service.title)}
+                                    itemComponent={ComboBoxListItem}
+                                    valueField="title"
+                                    textField="title"
+                                    filter="contains"
+                                    placeholder={LocaleUtils.getMessageById(this.context.messages, "catalog.servicePlaceholder")}
+                                    />
 
-                     {this.props.wrapOptions ? <Panel collapsible defaultExpanded={false} header={LocaleUtils.getMessageById(this.context.messages, "catalog.options")}>
-                         {textSearch}
-                     </Panel> : textSearch}
-                     {this.renderButtons()}
-                 </div>
-                 <div>
+                              <InputGroup.Addon onClick={() => this.props.onChangeCatalogMode("edit", this.isServiceEmpty())}>
+                                    <Glyphicon glyph={this.isServiceEmpty() ? "pencil-add" : "pencil"}/>
+                                </InputGroup.Addon>
+                            </InputGroup>
+                        </FormGroup>
+                        <FormGroup controlId="searchText" key="searchText">
+                            {this.renderTextSearch()}
+                        </FormGroup>
+                        <FormGroup controlId="buttons" key="buttons">
+                            {this.renderButtons()}
+                        </FormGroup>
+                    </Form>
+                ) : (
+                        <Form horizontal >
+                            <FormGroup>
+                                <Col xs={12} sm={3} md={3}>
+                                    <ControlLabel><Message msgId="catalog.type"/></ControlLabel>
+                                    <FormControl
+                                        onChange={(e) => this.props.onChangeNewType(e.target.value)}
+                                        value={this.props.newService && this.props.newService.type}
+                                        componentClass="select">
+                                        {this.renderFormats()}
+                                    </FormControl>
+                                </Col>
+                                <Col xs={12} sm={9} md={9}>
+                                    <ControlLabel><Message msgId="catalog.url"/></ControlLabel>
+                                <FormControl
+                                    ref="url"
+                                    type="text"
+                                    style={{
+                                        textOverflow: "ellipsis"
+                                    }}
+                                    placeholder={LocaleUtils.getMessageById(this.context.messages, "catalog.urlPlaceholder")}
+                                    value={this.props.newService && this.props.newService.url}
+                                    onChange={(e) => this.props.onChangeNewUrl(e.target.value)}
+                                    onKeyDown={this.onKeyDown}/>
+                                </Col>
+                            </FormGroup>
+                            <FormGroup controlId="title" key="title">
+                                <Col xs={12}>
+                                    <ControlLabel><Message msgId="catalog.serviceTitle"/></ControlLabel>
+                                <FormControl
+                                    ref="title"
+                                    type="text"
+                                    style={{
+                                        textOverflow: "ellipsis"
+                                    }}
+                                    placeholder={LocaleUtils.getMessageById(this.context.messages, "catalog.serviceTitlePlaceholder")}
+                                    value={this.props.newService && this.props.newService.title}
+                                    onChange={(e) => this.props.onChangeNewTitle(e.target.value)}
+                                    onKeyDown={this.onKeyDown}/>
+                                </Col>
+                            </FormGroup>
+                            <FormGroup controlId="buttons" key="butStons">
+                                <Col xs={12}>
+                                    {this.renderButtons()}
+                                </Col>
+                            </FormGroup>
+                        </Form>
+                    )}
+                    </div>
+                <div>
                     {this.renderResult()}
                     {this.props.layerError ? this.renderError(this.props.layerError) : null}
-                 </div>
-             </div>
+                </div>
+            </div>
         );
     }
 
+    isServiceEmpty = () => {
+        return this.props.selectedService === "";
+    };
+
+    isValidServiceSelected = () => {
+        return this.props.services[this.props.selectedService] !== undefined;
+    };
     search = () => {
-        this.props.onSearch(this.props.format, this.getCatalogUrl(), 1, this.props.pageSize, this.state && this.state.searchText);
+        const url = this.props.services[this.props.selectedService].url;
+        const type = this.props.services[this.props.selectedService].type;
+        this.props.onSearch(type, url, 1, this.props.pageSize, this.state && this.state.searchText || "");
         this.setState({
             loading: true
         });
     };
+
+    isViewMode = (mode) => {
+        return mode === "view";
+    }
 
     reset = () => {
         if (this.refs.catalogURL) {
