+++ web/client/translations/data.en-US
@@ -665,13 +665,19 @@
             "share": "Share",
             "copyToClipboard": "Copy to clipboard",
             "copied": "Copied!",
-            "catalogUrlPlaceholder": "Enter catalog URL...",
             "textSearchPlaceholder": "text to search...",
             "search": "Search",
             "reset": "Clear",
             "options": "Options",
             "srs_not_allowed": "The map coordinate system is not supported by this service",
             "add": "Add",
+            "service": "Service",
+            "servicePlaceholder": "type a service",
+            "url": "Url",
+            "urlPlaceholder": "type a URL",
+            "type": "type",
+            "serviceTitle": "Title",
+            "serviceTitlePlaceholder": "type a title",
             "notification": {
                 "warningAddCatalogService": "Insert a valid url and title",
                 "addCatalogService": "Service added correctly",
